work on logic
