+++ src/app/(misc)/currency/page.tsx
@@ -1,20 +1,15 @@
 "use client"
 
-import { useState } from "react"
 
-import PlaceBetComponent from "@/components/game-logic/b"
-import AmountMines from "@/components/settings/AmountMines"
+import AmountBombs from "@/components/settings/AmountBombs"
 import SelectTiles from "@/components/settings/AmountTiles"
 import BalanceBetSizeLogic from "@/components/settings/BalanceBetSizeLogic"
-import Minesweeper from "@/components/shells/v1"
-
 export default function CurrencyPage() {
-  const [bet, setBet] = useState(0)
   return (
     <main className="p-10 text-xl">
       <BalanceBetSizeLogic />
-      <SelectTiles />
-      {/* <AmountMines/> */}
+      <AmountBombs />
+      <SelectTiles/>
       {/* <PlaceBetComponent bet={bet} setBet={setBet} /> */}
       {/* <Minesweeper/> */}
     </main>
