+++ src/components/shells/v1.tsx
@@ -11,7 +11,7 @@ import { toast } from "sonner"
 import BetSize from "../game-logic/betSize"
 import FreezeGame from "../game-logic/freezeGame"
 import ToggleHoldMouse from "../game-logic/toggleHoldMouse"
-import SelectTiles from "../settings/AmountTiles"
+import SelectTiles from "../settings/AmountBombs"
 import BalanceBetSize from "../settings/BalanceBetSize"
 import SelectMode from "../settings/SelectGameMode"
 import StartGame from "../settings/StartGame"
